+++ src/main/java/com/triprecord/triprecord/global/exception/CustomJwtAuthenticationEntryPoint.java
